requirements.txt
# ez_setup
flask
# numpy
opencv-python
pillow
python-gflags
# setuptools
requests
tensorflow

# -e git://github.com/veegee82/tf_base.git#egg=tf_base